+++ crates/core/app/src/action_handler/actions.rs
@@ -3,8 +3,8 @@ use std::sync::Arc;
 use anyhow::Result;
 use async_trait::async_trait;
 use penumbra_chain::TransactionContext;
-use penumbra_ibc::component::transfer::Ics20Transfer;
 use penumbra_ibc::component::StateReadExt as _;
+use penumbra_shielded_pool::component::Ics20Transfer;
 use penumbra_storage::{StateRead, StateWrite};
 use penumbra_transaction::Action;
