+++ crates/core/component/ibc/src/component/action_handler.rs
@@ -1,2 +1,2 @@
 mod ibc_action;
-mod ics20_withdrawal;
+//mod ics20_withdrawal;
