+++ crates/core/component/shielded-pool/src/component.rs
@@ -5,10 +5,12 @@ mod metrics;
 mod note_manager;
 mod shielded_pool;
 mod supply;
+mod transfer;
 
 pub use self::metrics::register_metrics;
 pub use note_manager::NoteManager;
 pub use shielded_pool::{ShieldedPool, StateReadExt};
 pub use supply::{SupplyRead, SupplyWrite};
+pub use transfer::Ics20Transfer;
 
 pub mod rpc;
