+++ crates/core/component/shielded-pool/src/component/action_handler.rs
@@ -1,2 +1,3 @@
+mod ics20_withdrawal;
 mod output;
 mod spend;
