+++ crates/core/transaction/src/effect_hash.rs
@@ -11,7 +11,6 @@ use penumbra_governance::{
     DelegatorVote, DelegatorVoteBody, Proposal, ProposalDepositClaim, ProposalSubmit,
     ProposalWithdraw, ValidatorVote, ValidatorVoteBody, Vote,
 };
-use penumbra_ibc::Ics20Withdrawal;
 use penumbra_keys::{FullViewingKey, PayloadKey};
 use penumbra_proto::{
     core::component::dex::v1alpha1 as pbd, core::component::fee::v1alpha1 as pbf,
@@ -20,7 +19,7 @@ use penumbra_proto::{
     core::transaction::v1alpha1 as pbt, crypto::decaf377_fmd::v1alpha1 as pb_fmd, Message,
 };
 use penumbra_proto::{DomainType, TypeUrl};
-use penumbra_shielded_pool::{output, spend};
+use penumbra_shielded_pool::{output, spend, Ics20Withdrawal};
 use penumbra_stake::{Delegate, Undelegate, UndelegateClaimBody};
 
 use crate::{
