+++ crates/core/transaction/src/plan.rs
@@ -13,10 +13,10 @@ use penumbra_fee::Fee;
 use penumbra_governance::{
     DelegatorVotePlan, ProposalDepositClaim, ProposalSubmit, ProposalWithdraw, ValidatorVote,
 };
-use penumbra_ibc::{IbcAction, Ics20Withdrawal};
+use penumbra_ibc::IbcAction;
 use penumbra_keys::Address;
 use penumbra_proto::{core::transaction::v1alpha1 as pb, DomainType, TypeUrl};
-use penumbra_shielded_pool::{OutputPlan, SpendPlan};
+use penumbra_shielded_pool::{Ics20Withdrawal, OutputPlan, SpendPlan};
 use penumbra_stake::{Delegate, Undelegate, UndelegateClaimPlan};
 use rand::{CryptoRng, Rng};
 use serde::{Deserialize, Serialize};
